+++ wellregistry/wellregistry/urls.py
@@ -19,8 +19,11 @@ from django.urls import include, path
 
 urlpatterns = [
     # this is the django admin url which allows adding django users and table management
-    path('admin/', admin.site.urls),
+    path('registry/admin/', admin.site.urls),
+
+    path('', include('social_django.urls', namespace='social')),
+    path('registry/accounts/', include('django.contrib.auth.urls')),
 
     # this is our registry page
-    path('registry/', include('registry.urls'))
+    path('', include('registry.urls')),
 ]