+++ CHANGELOG.md
@@ -7,4 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##[Unreleased](https://github.com/ACWI-SOGW/well_registry_management/tree/master)
 ### Added
-- Added Registry model and admin prototype interface.
+-   Added Registry model and admin prototype interface.
+-   Added initial registry model with migrations and admin 
+-   Added keycloak as the authentication method for the admin.